sayt/requirements.txt
djangorestframework==3.14.0
    Django==4.1.7
        asgiref==3.6.0
        sqlparse==0.4.3
        tzdata==2022.7
    pytz==2022.7.1
pipfreeze==2.0.3
